Fix strnicmp not defined in Kernel >= 4.0
This particular fix is all over the internet, as patches posted on
mailing lists, package maintainer's mirrors, etc.

I don't know the full origin of it, but it works. It also makes sense. I
found a quote regarding it here:

http://comments.gmane.org/gmane.linux.drivers.driver-project.devel/71438
> [...] patched so it will compile against kernels that no longer
> provide the old procfs API and with calls to strnicmp() replaced with
> strncasecmp().

Here's an example of the patch in the wild:
https://github.com/felixonmars/aur3-mirror/blob/9cf1f08f9a6806ae5804f534bceb174cfec3daa9/8192cu-dkms/RTL8192CU-kernel-4.0.patch

Note that the sources mentioned above all talk about the more common
8192cu driver, but the change is the same as both drivers share the
majority of the code with each other.
<speculation>
I guess this is part of the reason why the hardware makers don't publish
updated drivers; they've copy-pasted it everywhere and patching them all
is a pain for them.
</speculation>